Add files via upload
pyEVP.py is still in the process of being written. I'm trying not to make too many changes to the structure yet; the goal is to eventually turn most of the functionality of the program into functions and objects, which can more easily be re-used and scripted. Another option is to keep much of the program in FORTRAN (for computational speed), but rewritten as functions which can then be called through a main program in python.